Fix yet another bug with crashing
Turns out, as the proposer, we *must* save the last proposal we made.
Sending out duplicate Prepares is okay, but we must never send out
duplicate Accepts, and to avoid that, we gotta track which proposals
we've made.

This surfaced from the check in the learner that confirms that the
values we get for a given proposal are consistent. I thought this was
because we were sending too many Accepts in response to Promises, but
turns out, it's much trickier: after sending out an Accept(0, ...), we
crashed, then sent another Prepare(0) request. The second one hit a
different set of Acceptors, so when we sent out a second Accept(0, ...),
we sent out a different value.

I don't like this solution but it does seem to work. I knew distributed
systems were hard, but this really exceeded expectations.